+++ modules/table/src/test/java/org/apache/ignite/internal/table/distributed/IndexCleanupTest.java
@@ -17,144 +17,30 @@
 
 package org.apache.ignite.internal.table.distributed;
 
-import static java.util.Collections.singletonMap;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.contains;
 import static org.hamcrest.Matchers.not;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import java.util.List;
-import java.util.Map;
 import java.util.Set;
 import java.util.UUID;
-import org.apache.ignite.distributed.TestPartitionDataStorage;
-import org.apache.ignite.internal.hlc.HybridClock;
-import org.apache.ignite.internal.hlc.HybridClockImpl;
 import org.apache.ignite.internal.schema.BinaryRow;
-import org.apache.ignite.internal.schema.BinaryRowConverter;
-import org.apache.ignite.internal.schema.BinaryTupleSchema;
-import org.apache.ignite.internal.schema.Column;
-import org.apache.ignite.internal.schema.NativeTypes;
-import org.apache.ignite.internal.schema.SchemaDescriptor;
-import org.apache.ignite.internal.schema.marshaller.KvMarshaller;
-import org.apache.ignite.internal.schema.marshaller.MarshallerException;
-import org.apache.ignite.internal.schema.marshaller.MarshallerFactory;
-import org.apache.ignite.internal.schema.marshaller.reflection.ReflectionMarshallerFactory;
 import org.apache.ignite.internal.storage.RowId;
-import org.apache.ignite.internal.storage.impl.TestMvPartitionStorage;
-import org.apache.ignite.internal.storage.index.HashIndexDescriptor;
-import org.apache.ignite.internal.storage.index.HashIndexDescriptor.HashIndexColumnDescriptor;
-import org.apache.ignite.internal.storage.index.SortedIndexDescriptor;
-import org.apache.ignite.internal.storage.index.SortedIndexDescriptor.SortedIndexColumnDescriptor;
-import org.apache.ignite.internal.storage.index.impl.TestHashIndexStorage;
-import org.apache.ignite.internal.storage.index.impl.TestSortedIndexStorage;
-import org.apache.ignite.internal.table.distributed.replicator.TablePartitionId;
-import org.jetbrains.annotations.Nullable;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
 
 /** Tests indexes cleaning up on data removal and transaction abortions. */
-public class IndexCleanupTest {
-    /** Default reflection marshaller factory. */
-    private static final MarshallerFactory MARSHALLER_FACTORY = new ReflectionMarshallerFactory();
-
-    private static final SchemaDescriptor SCHEMA_DESCRIPTOR = new SchemaDescriptor(1, new Column[]{
-            new Column("INTKEY", NativeTypes.INT32, false),
-            new Column("STRKEY", NativeTypes.STRING, false),
-    }, new Column[]{
-            new Column("INTVAL", NativeTypes.INT32, false),
-            new Column("STRVAL", NativeTypes.STRING, false),
-    });
-
-    private static final BinaryTupleSchema TUPLE_SCHEMA = BinaryTupleSchema.createRowSchema(SCHEMA_DESCRIPTOR);
-
-    private static final BinaryTupleSchema PK_INDEX_SCHEMA = BinaryTupleSchema.createKeySchema(SCHEMA_DESCRIPTOR);
-
-    private static final BinaryRowConverter PK_INDEX_BINARY_TUPLE_CONVERTER = new BinaryRowConverter(TUPLE_SCHEMA, PK_INDEX_SCHEMA);
-
-    private static final int[] USER_INDEX_COLS = {
-            SCHEMA_DESCRIPTOR.column("INTVAL").schemaIndex(),
-            SCHEMA_DESCRIPTOR.column("STRVAL").schemaIndex()
-    };
-
-    private static final BinaryTupleSchema USER_INDEX_SCHEMA = BinaryTupleSchema.createSchema(SCHEMA_DESCRIPTOR, USER_INDEX_COLS);
-
-    private static final BinaryRowConverter USER_INDEX_BINARY_TUPLE_CONVERTER = new BinaryRowConverter(TUPLE_SCHEMA, USER_INDEX_SCHEMA);
-
-    /** Key-value marshaller for tests. */
-    private static final KvMarshaller<TestKey, TestValue> KV_MARSHALLER
-            = MARSHALLER_FACTORY.create(SCHEMA_DESCRIPTOR, TestKey.class, TestValue.class);
-
-    private static final UUID TX_ID = UUID.randomUUID();
-
-    private static final HybridClock CLOCK = new HybridClockImpl();
-
-    private TestHashIndexStorage pkInnerStorage;
-    private TestSortedIndexStorage sortedInnerStorage;
-    private TestHashIndexStorage hashInnerStorage;
-    private TestMvPartitionStorage storage;
-    private StorageUpdateHandler storageUpdateHandler;
-
-    @BeforeEach
-    void setUp() {
-        UUID pkIndexId = UUID.randomUUID();
-        UUID sortedIndexId = UUID.randomUUID();
-        UUID hashIndexId = UUID.randomUUID();
-
-        pkInnerStorage = new TestHashIndexStorage(null);
-
-        TableSchemaAwareIndexStorage pkStorage = new TableSchemaAwareIndexStorage(
-                pkIndexId,
-                pkInnerStorage,
-                PK_INDEX_BINARY_TUPLE_CONVERTER::toTuple
-        );
-
-        sortedInnerStorage = new TestSortedIndexStorage(new SortedIndexDescriptor(sortedIndexId, List.of(
-                new SortedIndexColumnDescriptor("INTVAL", NativeTypes.INT32, false, true),
-                new SortedIndexColumnDescriptor("STRVAL", NativeTypes.STRING, false, true)
-        )));
-
-        TableSchemaAwareIndexStorage sortedIndexStorage = new TableSchemaAwareIndexStorage(
-                sortedIndexId,
-                sortedInnerStorage,
-                USER_INDEX_BINARY_TUPLE_CONVERTER::toTuple
-        );
-
-        hashInnerStorage = new TestHashIndexStorage(new HashIndexDescriptor(hashIndexId, List.of(
-                new HashIndexColumnDescriptor("INTVAL", NativeTypes.INT32, false),
-                new HashIndexColumnDescriptor("STRVAL", NativeTypes.STRING, false)
-        )));
-
-        TableSchemaAwareIndexStorage hashIndexStorage = new TableSchemaAwareIndexStorage(
-                hashIndexId,
-                hashInnerStorage,
-                USER_INDEX_BINARY_TUPLE_CONVERTER::toTuple
-        );
-
-        storage = new TestMvPartitionStorage(1);
-
-        storageUpdateHandler = new StorageUpdateHandler(1, new TestPartitionDataStorage(storage),
-                () -> Map.of(
-                        pkIndexId, pkStorage,
-                        sortedIndexId, sortedIndexStorage,
-                        hashIndexId, hashIndexStorage
-                )
-        );
-    }
-
+public class IndexCleanupTest extends IndexBaseTest {
     @ParameterizedTest
     @EnumSource(AddWrite.class)
     void testAbort(AddWrite writer) {
         UUID rowUuid = UUID.randomUUID();
         RowId rowId = new RowId(1, rowUuid);
 
-        var key = new TestKey(1, "foo");
-        var value = new TestValue(2, "bar");
-        BinaryRow tableRow = binaryRow(key, value);
+        BinaryRow row = defaultRow();
 
-        writer.addWrite(storageUpdateHandler, rowUuid, tableRow);
+        writer.addWrite(storageUpdateHandler, rowUuid, row);
 
         assertEquals(1, storage.rowsCount());
         assertThat(pkInnerStorage.allRowsIds(), contains(rowId));
@@ -175,11 +61,9 @@ public class IndexCleanupTest {
         UUID rowUuid = UUID.randomUUID();
         RowId rowId = new RowId(1, rowUuid);
 
-        var key = new TestKey(1, "foo");
-        var value = new TestValue(2, "bar");
-        BinaryRow tableRow = binaryRow(key, value);
+        BinaryRow row = defaultRow();
 
-        writer.addWrite(storageUpdateHandler, rowUuid, tableRow);
+        writer.addWrite(storageUpdateHandler, rowUuid, row);
 
         // Write intent is in the storage.
         assertEquals(1, storage.rowsCount());
@@ -206,11 +90,9 @@ public class IndexCleanupTest {
         UUID rowUuid = UUID.randomUUID();
         RowId rowId = new RowId(1, rowUuid);
 
-        var key = new TestKey(1, "foo");
-        var value = new TestValue(2, "bar");
-        BinaryRow tableRow = binaryRow(key, value);
+        BinaryRow row = defaultRow();
 
-        writer.addWrite(storageUpdateHandler, rowUuid, tableRow);
+        writer.addWrite(storageUpdateHandler, rowUuid, row);
         writer.addWrite(storageUpdateHandler, rowUuid, null);
 
         storageUpdateHandler.handleTransactionAbortion(Set.of(rowId), () -> {});
@@ -261,11 +143,9 @@ public class IndexCleanupTest {
         UUID rowUuid = UUID.randomUUID();
         RowId rowId = new RowId(1, rowUuid);
 
-        var key = new TestKey(1, "foo");
-        var value = new TestValue(2, "bar");
-        BinaryRow tableRow = binaryRow(key, value);
+        BinaryRow row = defaultRow();
 
-        writer.addWrite(storageUpdateHandler, rowUuid, tableRow);
+        writer.addWrite(storageUpdateHandler, rowUuid, row);
         commitWrite(rowId);
 
         writer.addWrite(storageUpdateHandler, rowUuid, null);
@@ -284,14 +164,12 @@ public class IndexCleanupTest {
         UUID rowUuid = UUID.randomUUID();
         RowId rowId = new RowId(1, rowUuid);
 
-        var key = new TestKey(1, "foo");
-        var value = new TestValue(2, "bar");
-        BinaryRow tableRow = binaryRow(key, value);
+        BinaryRow row = defaultRow();
 
-        writer.addWrite(storageUpdateHandler, rowUuid, tableRow);
+        writer.addWrite(storageUpdateHandler, rowUuid, row);
         commitWrite(rowId);
 
-        writer.addWrite(storageUpdateHandler, rowUuid, tableRow);
+        writer.addWrite(storageUpdateHandler, rowUuid, row);
 
         storageUpdateHandler.handleTransactionAbortion(Set.of(rowId), () -> {});
 
@@ -307,14 +185,12 @@ public class IndexCleanupTest {
         UUID rowUuid = UUID.randomUUID();
         RowId rowId = new RowId(1, rowUuid);
 
-        var key = new TestKey(1, "foo");
-        var value = new TestValue(2, "bar");
-        BinaryRow tableRow = binaryRow(key, value);
+        BinaryRow row = defaultRow();
 
-        writer.addWrite(storageUpdateHandler, rowUuid, tableRow);
+        writer.addWrite(storageUpdateHandler, rowUuid, row);
         commitWrite(rowId);
 
-        writer.addWrite(storageUpdateHandler, rowUuid, tableRow);
+        writer.addWrite(storageUpdateHandler, rowUuid, row);
         writer.addWrite(storageUpdateHandler, rowUuid, null);
 
         assertEquals(1, storage.rowsCount());
@@ -329,103 +205,20 @@ public class IndexCleanupTest {
         UUID rowUuid = UUID.randomUUID();
         RowId rowId = new RowId(1, rowUuid);
 
-        var key = new TestKey(1, "foo");
-        var value = new TestValue(2, "bar");
-        BinaryRow tableRow = binaryRow(key, value);
+        BinaryRow row = defaultRow();
 
-        writer.addWrite(storageUpdateHandler, rowUuid, tableRow);
+        writer.addWrite(storageUpdateHandler, rowUuid, row);
         commitWrite(rowId);
 
-        writer.addWrite(storageUpdateHandler, rowUuid, tableRow);
+        writer.addWrite(storageUpdateHandler, rowUuid, row);
         commitWrite(rowId);
 
-        writer.addWrite(storageUpdateHandler, rowUuid, tableRow);
+        writer.addWrite(storageUpdateHandler, rowUuid, row);
         writer.addWrite(storageUpdateHandler, rowUuid, null);
 
         assertEquals(1, storage.rowsCount());
         assertThat(pkInnerStorage.allRowsIds(), contains(rowId));
         assertThat(sortedInnerStorage.allRowsIds(), contains(rowId));
         assertThat(hashInnerStorage.allRowsIds(), contains(rowId));
-    }
-
-    /** Enum that encapsulates update API. */
-    enum AddWrite {
-        /** Uses update api. */
-        USE_UPDATE {
-            @Override
-            void addWrite(StorageUpdateHandler handler, TablePartitionId partitionId, UUID rowUuid, @Nullable BinaryRow row) {
-                handler.handleUpdate(
-                        TX_ID,
-                        rowUuid,
-                        partitionId,
-                        row == null ? null : row.byteBuffer(),
-                        (unused) -> {}
-                );
-            }
-        },
-        /** Uses updateAll api. */
-        USE_UPDATE_ALL {
-            @Override
-            void addWrite(StorageUpdateHandler handler, TablePartitionId partitionId, UUID rowUuid, @Nullable BinaryRow row) {
-                handler.handleUpdateAll(
-                        TX_ID,
-                        singletonMap(rowUuid, row == null ? null : row.byteBuffer()),
-                        partitionId,
-                        (unused) -> {}
-                );
-            }
-        };
-
-        void addWrite(StorageUpdateHandler handler, UUID rowUuid, @Nullable BinaryRow row) {
-            TablePartitionId tablePartitionId = new TablePartitionId(UUID.randomUUID(), 1);
-
-            addWrite(handler, tablePartitionId, rowUuid, row);
-        }
-
-        abstract void addWrite(StorageUpdateHandler handler, TablePartitionId partitionId, UUID rowUuid, @Nullable BinaryRow row);
-    }
-
-    private static BinaryRow binaryRow(TestKey key, TestValue value) {
-        try {
-            return KV_MARSHALLER.marshal(key, value);
-        } catch (MarshallerException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    private void commitWrite(RowId rowId) {
-        storage.runConsistently(() -> {
-            storage.commitWrite(rowId, CLOCK.now());
-
-            return null;
-        });
-    }
-
-    private static class TestKey {
-        int intKey;
-
-        String strKey;
-
-        TestKey() {
-        }
-
-        TestKey(int intKey, String strKey) {
-            this.intKey = intKey;
-            this.strKey = strKey;
-        }
-    }
-
-    private static class TestValue {
-        Integer intVal;
-
-        String strVal;
-
-        TestValue() {
-        }
-
-        TestValue(Integer intVal, String strVal) {
-            this.intVal = intVal;
-            this.strVal = strVal;
-        }
     }
 }
